Update README.md
NFT Tech World was started with a dream to provide world-class online solutions to businesses around the world. NFTTW has proliferated globally and now has become a renowned name, primarily due to its powerhouse combination of qualified professionals, latest technologies and tenacious passion for what we do. NFT TECH World has a rich expertise in Mobile Development, Web Development, E-commerce, Blockchain Development and Digital Marketing. We also specialize in On-Demand Mobile Application Development and the various sectors of Blockchain Development such as DApp, Cryptocurrency development, ICO development and Marketing, Smart Contract Development and Hyperledger development. Furthermore, we offer a platform to Hire Dedicated Developers on a weekly/monthly contract model based on the client’s needs. Visit nfttechworld.io! Our passionate developers embrace new technologies to build applications and leverage them in order to deliver results that would be beneficial to our clients not only for the present but also in the days to come. Our strategies and ventures prove our skills and proficiency in providing top-notch results to gratify our clients. We render astute services and make every effort to spread a positive impact across the world. Tailored to meet global standards, our services are employed in a seamless work flow, with strict adherence to quality. We provide enterprise solutions that are within your budget, fall in line with your business needs and are delivered on time. We approach every client with a view of building long-term partnership. We focus constantly on meeting clients’ expectation and delivering solutions with absolute transparency. We hold honesty and integrity with high regard. We only promise what we can deliver and not add features that will just end up costing you more. We stick to no cookie cutter practices. Our approach is creative and innovative. With our world-class services, we deliver ethical, quality driven and cost-effective solutions.